Fixing "function might not be defined at runtime" errors.
Without this, I get the following errors when depending on this package:

Error: the function "benchmark-init/node-root-p" might not be defined at runtime.
Error: the function "benchmark-init/node-duration-adjusted" might not be defined at runtime.
Error: the function "benchmark-init/flatten" might not be defined at runtime.

This error is accurate as the eval-when-compile does not force the
require during a normal load.